+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Add support for rolling cult events with descending DC
 * Add support for tracking and gaining XP from cult events that have already occurred
+* Use built in V11 weather effects instead of FXMaster; you need to manually clear existing weather in FXMaster
 
 ## [0.1.0] - 2023-06-11